requirements.txt
apturl==0.5.2
asn1crypto==0.24.0
Brlapi==0.6.6
certifi==2018.10.15
chardet==3.0.4
clearbit==0.1.7
command-not-found==0.3
cryptography==2.1.4
cupshelpers==1.0
defer==1.0.6
defusedxml==0.5.0
distro-info==0.18
Django==2.1.5
django-allauth==0.38.0
django-cors-headers==2.4.0
django-filter==2.0.0
django-rest-auth==0.9.3
django-rest-framework==0.1.0
djangorestframework==3.9.1
djangorestframework-jwt==1.11.0
httplib2==0.11.3
idna==2.7
keyring==10.6.0
keyrings.alt==3.0
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
louis==3.5.0
macaroonbakery==1.1.3
Mako==1.0.7
Markdown==3.0.1
MarkupSafe==1.0
mysqlclient==1.3.14
netifaces==0.10.4
oauth==1.0.1
oauthlib==2.1.0
olefile==0.45.1
pexpect==4.2.1
Pillow==5.1.0
protobuf==3.0.0
pycairo==1.16.2
pycrypto==2.6.1
pycups==1.9.73
pygobject==3.26.1
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.1.2
pyRFC3339==1.0
python-apt==1.6.3
python-debian==0.1.32
python3-openid==3.1.0
pytz==2018.9
pyxdg==0.25
PyYAML==3.12
reportlab==3.4.0
requests==2.20.0
requests-oauthlib==1.0.0
requests-unixsocket==0.1.5
screen-resolution-extra==0.0.0
SecretStorage==2.3.1
simplejson==3.13.2
six==1.11.0
speedtest-cli==2.0.2
system-service==0.3
systemd-python==234
ubuntu-drivers-common==0.0.0
ufw==0.35
unattended-upgrades==0.1
urllib3==1.24
usb-creator==0.3.3
virtualenv==16.2.0
wadllib==1.3.2
xkit==0.0.0
zope.interface==4.3.2
